+++ src/Data/Either/Validation.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE Rank2Types #-}
+
 -----------------------------------------------------------------------------
 -- |
 -- Module      :  Data.Either.Validation
@@ -14,6 +16,11 @@
 
 module Data.Either.Validation
   ( Validation(..)
+  , _Success
+  , _Failure
+  , eitherToValidation
+  , validationToEither
+  , _Validation
   ) where
 
 import Control.Applicative
@@ -23,6 +30,7 @@ import Data.Bitraversable(Bitraversable(bitraverse))
 import Data.Foldable (Foldable(foldr))
 import Data.Functor.Alt (Alt((<!>)))
 import Data.Monoid (Monoid(mappend, mempty))
+import Data.Profunctor
 import Data.Semigroup (Semigroup((<>)))
 import Data.Traversable (Traversable(traverse))
 import Prelude hiding (foldr)
@@ -84,3 +92,50 @@ instance Monoid e => Monoid (Validation e a) where
   Failure _  `mappend` Success a2 = Success a2
   Success a1 `mappend` Failure _  = Success a1
   Success a1 `mappend` Success _  = Success a1
+
+type Prism s t a b = forall p f. (Choice p, Applicative f) => p a (f b) -> p s (f t)
+
+prism :: (b -> t) -> (s -> Either t a) -> Prism s t a b
+prism bt seta = dimap seta (either pure (fmap bt)) . right'
+{-# INLINE prism #-}
+
+_Failure :: Prism (Validation a c) (Validation b c) a b
+_Failure = prism
+           (\ x -> Failure x)
+           (\ x
+            -> case x of
+              Failure y -> Right y
+              Success y -> Left (Success y))
+{-# INLINE _Failure #-}
+
+_Success :: Prism (Validation c a) (Validation c b) a b
+_Success = prism
+           (\ x -> Success x)
+           (\ x
+            -> case x of
+              Failure y -> Left (Failure y)
+              Success y -> Right y)
+{-# INLINE _Success #-}
+
+type Iso s t a b = (Profunctor p, Functor f) => p a (f b) -> p s (f t)
+
+iso :: (s -> a) -> (b -> t) -> Iso s t a b
+iso sa bt = dimap sa (fmap bt)
+{-# INLINE iso #-}
+
+validationToEither :: Validation e a -> Either e a
+validationToEither x = case x of
+  Failure e -> Left e
+  Success a -> Right a
+{-# INLINE validationToEither #-}
+
+eitherToValidation :: Either e a -> Validation e a
+eitherToValidation x = case x of
+  Left e -> Failure e
+  Right a -> Success a
+{-# INLINE eitherToValidation #-}
+
+-- | 'Validation' is isomorphic to 'Either'
+_Validation :: Iso (Validation e a) (Validation g b) (Either e a) (Either g b)
+_Validation = iso validationToEither eitherToValidation
+{-# INLINE _Validation #-}